Formatting updates and moving markov quote generator off into separate module for future extraction into dedicated API
