+++ README.md
@@ -5,8 +5,13 @@ All you should need is a `bot.py` config file that looks like this:
 ```
 CLIENT_ID=''
 CLIENT_SECRET=''
+DISCORD_TOKEN=''
+FOOTER=''
 PASSWORD=''
+STATE_SIZE=
+SUB=''
 USERNAME=''
 USER_AGENT=''
-DISCORD_TOKEN=''
 ```
+
+STATE_SIZE is recommended to be set to 2
